Moving precondition.ts back to where it was originally
