+++ AzureFunctions.AngularClient/src/app/shared/preconditions.ts
@@ -1,12 +1,12 @@
 import { CacheService } from 'app/shared/services/cache.service';
 import { Observable } from 'rxjs/Observable';
-import { ArmObj } from '../models/arm/arm-obj';
-import { Site } from '../models/arm/site';
-import { errorIds } from '../models/error-ids';
-import { HostingEnvironment } from '../models/arm/hosting-environment';
-import { LogService } from '../services/log.service';
+import { ArmObj } from './models/arm/arm-obj';
+import { Site } from './models/arm/site';
+import { errorIds } from './models/error-ids';
+import { HostingEnvironment } from './models/arm/hosting-environment';
+import { LogService } from './services/log.service';
 import { Headers } from '@angular/http';
-import { HostStatus } from '../models/host-status';
+import { HostStatus } from './models/host-status';
 import { Injector } from '@angular/core';
 import { FunctionAppContext } from 'app/shared/function-app-context';
 import { ArmUtil } from 'app/shared/Utilities/arm-utils';
